+++ tokio-buf/Cargo.toml
@@ -20,3 +20,6 @@ categories = ["asynchronous"]
 bytes = { version = "0.4.10", features = [ "either" ] }
 either = "1.5"
 futures = "0.1.23"
+
+[features]
+core = []
