+++ tokio-buf/src/ext/bytes.rs
@@ -1,5 +1,5 @@
 use BufStream;
-use buf_stream::errors::internal::Never;
+use ext::errors::internal::Never;
 
 use bytes::{Bytes, BytesMut};
 use futures::Poll;
