+++ tokio-buf/src/ext/from.rs
@@ -1,4 +1,4 @@
-use super::SizeHint;
+use SizeHint;
 
 use bytes::{Buf, BufMut};
