+++ tokio-buf/src/ext/limit.rs
@@ -1,4 +1,5 @@
-use super::{BufStream, SizeHint};
+use BufStream;
+use super::BufStreamExt;
 
 use bytes::Buf;
 use futures::Poll;
@@ -28,7 +29,7 @@ impl<T> Limit<T> {
 
 impl<T> BufStream for Limit<T>
 where
-    T: BufStream,
+    T: BufStreamExt,
 {
     type Item = T::Item;
     type Error = LimitError<T::Error>;
@@ -58,22 +59,6 @@ where
         }
 
         res
-    }
-
-    fn size_hint(&self) -> SizeHint {
-        let mut hint = self.stream.size_hint();
-
-        let upper = hint.upper()
-            .map(|upper| upper.min(self.remaining))
-            .unwrap_or(self.remaining);
-
-        hint.set_upper(upper);
-        hint
-    }
-
-    fn consume_hint(&mut self, amount: usize) {
-        // TODO: Should this be capped by `self.remaining`?
-        self.stream.consume_hint(amount)
     }
 }
