+++ tokio-buf/src/ext/str.rs
@@ -1,5 +1,5 @@
 use BufStream;
-use buf_stream::errors::internal::Never;
+use ext::errors::internal::Never;
 
 use futures::Poll;
